add PreprocessException pass
Add __FILE__ and __LINE__ arguments to the Exception-derived
construction call in the PreprocessException pass instead of
doing it in the lexer.

The problem with lexer is that it tries to know about the
Exception class and adds these arguments too early.

If Exception class can be extended, then we don't know
whether a `new T()` expression constructs an exception class.
The PreprocessException has class data info, so it can
perform this operation with more accuracy.

We also change the Exception class constructor signature so
it's compatible with the PHP version. It doesn't accept the
file and line arguments now; to set these fields, a new builtin function
is added that is inserted by the compiler.

This transformation can be described as:

        new T(args...)
        =>
        __exception_set_location(new T(args...), __FILE__, __LINE__)

Where __exception_set_location() call returns the modified exception.

This patch does not add missing Exception-related features,
but it makes it's needed to make the work in that direction possible.